+++ js/packages/cli/src/helpers/cache.ts
@@ -1,24 +1,33 @@
+import fs from 'fs';
 import path from 'path';
 import { CACHE_PATH } from './constants';
-import fs from 'fs';
 
 export function cachePath(
   env: string,
   cacheName: string,
   cPath: string = CACHE_PATH,
+  legacy: boolean = false,
 ) {
-  return path.join(cPath, `${env}-${cacheName}`);
+  const filename = `${env}-${cacheName}`;
+  return path.join(cPath, legacy ? filename : `${filename}.json`);
 }
 
 export function loadCache(
   cacheName: string,
   env: string,
   cPath: string = CACHE_PATH,
+  legacy: boolean = false,
 ) {
-  const path = cachePath(env, cacheName, cPath);
-  return fs.existsSync(path)
-    ? JSON.parse(fs.readFileSync(path).toString())
-    : undefined;
+  const path = cachePath(env, cacheName, cPath, legacy);
+
+  if (!fs.existsSync(path)) {
+    if (!legacy) {
+      return loadCache(cacheName, env, cPath, true);
+    }
+    return undefined;
+  }
+
+  return JSON.parse(fs.readFileSync(path).toString());
 }
 
 export function saveCache(