+++ js/packages/cli/src/helpers/various.ts
@@ -2,11 +2,195 @@ import { LAMPORTS_PER_SOL, AccountInfo } from '@solana/web3.js';
 import fs from 'fs';
 import weighted from 'weighted';
 import path from 'path';
-import { Program, web3 } from '@project-serum/anchor';
+import { BN, Program, web3 } from '@project-serum/anchor';
 import { Token, TOKEN_PROGRAM_ID } from '@solana/spl-token';
+import { StorageType } from './storage-type';
+import { getAtaForMint } from './accounts';
+import { CLUSTERS, DEFAULT_CLUSTER } from './constants';
 
 const { readFile } = fs.promises;
 
+export async function getCandyMachineV2Config(
+  walletKeyPair: web3.Keypair,
+  anchorProgram: Program,
+  configPath: any,
+): Promise<{
+  storage: StorageType;
+  ipfsInfuraProjectId: string;
+  number: number;
+  ipfsInfuraSecret: string;
+  awsS3Bucket: string;
+  retainAuthority: boolean;
+  mutable: boolean;
+  batchSize: number;
+  price: BN;
+  treasuryWallet: web3.PublicKey;
+  splToken: web3.PublicKey | null;
+  gatekeeper: null | {
+    expireOnUse: boolean;
+    gatekeeperNetwork: web3.PublicKey;
+  };
+  endSettings: null | [number, BN];
+  whitelistMintSettings: null | {
+    mode: any;
+    mint: web3.PublicKey;
+    presale: boolean;
+    discountPrice: null | BN;
+  };
+  hiddenSettings: null | {
+    name: string;
+    uri: string;
+    hash: Uint8Array;
+  };
+  goLiveDate: BN | null;
+  uuid: string;
+  arweaveJwk: string;
+}> {
+  const configString = fs.readFileSync(configPath);
+
+  //@ts-ignore
+  const config = JSON.parse(configString);
+
+  const {
+    storage,
+    ipfsInfuraProjectId,
+    number,
+    ipfsInfuraSecret,
+    awsS3Bucket,
+    noRetainAuthority,
+    noMutable,
+    batchSize,
+    price,
+    splToken,
+    splTokenAccount,
+    solTreasuryAccount,
+    gatekeeper,
+    endSettings,
+    hiddenSettings,
+    whitelistMintSettings,
+    goLiveDate,
+    uuid,
+    arweaveJwk,
+  } = config;
+
+  let wallet;
+  let parsedPrice = price;
+
+  const splTokenAccountFigured = splTokenAccount
+    ? splTokenAccount
+    : splToken
+    ? (
+        await getAtaForMint(
+          new web3.PublicKey(splToken),
+          walletKeyPair.publicKey,
+        )
+      )[0]
+    : null;
+  if (splToken) {
+    if (solTreasuryAccount) {
+      throw new Error(
+        'If spl-token-account or spl-token is set then sol-treasury-account cannot be set',
+      );
+    }
+    if (!splToken) {
+      throw new Error(
+        'If spl-token-account is set, spl-token must also be set',
+      );
+    }
+    const splTokenKey = new web3.PublicKey(splToken);
+    const splTokenAccountKey = new web3.PublicKey(splTokenAccountFigured);
+    if (!splTokenAccountFigured) {
+      throw new Error(
+        'If spl-token is set, spl-token-account must also be set',
+      );
+    }
+
+    const token = new Token(
+      anchorProgram.provider.connection,
+      splTokenKey,
+      TOKEN_PROGRAM_ID,
+      walletKeyPair,
+    );
+
+    const mintInfo = await token.getMintInfo();
+    if (!mintInfo.isInitialized) {
+      throw new Error(`The specified spl-token is not initialized`);
+    }
+    const tokenAccount = await token.getAccountInfo(splTokenAccountKey);
+    if (!tokenAccount.isInitialized) {
+      throw new Error(`The specified spl-token-account is not initialized`);
+    }
+    if (!tokenAccount.mint.equals(splTokenKey)) {
+      throw new Error(
+        `The spl-token-account's mint (${tokenAccount.mint.toString()}) does not match specified spl-token ${splTokenKey.toString()}`,
+      );
+    }
+
+    wallet = new web3.PublicKey(splTokenAccountKey);
+    parsedPrice = price * 10 ** mintInfo.decimals;
+    if (whitelistMintSettings?.discountPrice) {
+      whitelistMintSettings.discountPrice *= 10 ** mintInfo.decimals;
+    }
+  } else {
+    parsedPrice = price * 10 ** 9;
+    if (whitelistMintSettings?.discountPrice) {
+      whitelistMintSettings.discountPrice *= 10 ** 9;
+    }
+    wallet = solTreasuryAccount
+      ? new web3.PublicKey(solTreasuryAccount)
+      : walletKeyPair.publicKey;
+  }
+
+  if (whitelistMintSettings) {
+    whitelistMintSettings.mint = new web3.PublicKey(whitelistMintSettings.mint);
+    if (whitelistMintSettings?.discountPrice) {
+      whitelistMintSettings.discountPrice = new BN(
+        whitelistMintSettings.discountPrice,
+      );
+    }
+  }
+
+  if (endSettings) {
+    if (endSettings.endSettingType.date) {
+      endSettings.number = new BN(parseDate(endSettings.value));
+    } else if (endSettings.endSettingType.amount) {
+      endSettings.number = new BN(endSettings.value);
+    }
+    delete endSettings.value;
+  }
+
+  if (hiddenSettings) {
+    const utf8Encode = new TextEncoder();
+    hiddenSettings.hash = utf8Encode.encode(hiddenSettings.hash);
+  }
+
+  if (gatekeeper) {
+    gatekeeper.gatekeeperNetwork = new web3.PublicKey(
+      gatekeeper.gatekeeperNetwork,
+    );
+  }
+
+  return {
+    storage,
+    ipfsInfuraProjectId,
+    number,
+    ipfsInfuraSecret,
+    awsS3Bucket,
+    retainAuthority: !noRetainAuthority,
+    mutable: !noMutable,
+    batchSize,
+    price: new BN(parsedPrice),
+    treasuryWallet: wallet,
+    splToken: splToken ? new web3.PublicKey(splToken) : null,
+    gatekeeper,
+    endSettings,
+    hiddenSettings,
+    whitelistMintSettings,
+    goLiveDate: goLiveDate ? new BN(parseDate(goLiveDate)) : null,
+    uuid,
+    arweaveJwk,
+  };
+}
 export async function readJsonFile(fileName: string) {
   const file = await readFile(fileName, 'utf-8');
   return JSON.parse(file);
@@ -312,3 +496,12 @@ export const getPriceWithMantissa = async (
 
   return Math.ceil(price * mantissa);
 };
+
+export function getCluster(name: string): string {
+  for (const cluster of CLUSTERS) {
+    if (cluster.name === name) {
+      return cluster.url;
+    }
+  }
+  return DEFAULT_CLUSTER.url;
+}