+++ js/packages/common/src/contexts/connection.tsx
@@ -32,8 +32,6 @@ interface BlockhashAndFeeCalculator {
 
 export type ENDPOINT_NAME =
   | 'mainnet-beta'
-  | 'mainnet-beta-solana'
-  | 'mainnet-beta-serum'
   | 'testnet'
   | 'devnet'
   | 'localnet'
@@ -45,25 +43,12 @@ type Endpoint = {
   url: string;
   chainId: ChainId;
 };
-const LOCALHOST = 'localhost';
 
 export const ENDPOINTS: Array<Endpoint> = [
   {
     name: 'mainnet-beta',
     label: 'mainnet-beta',
     url: 'https://api.metaplex.solana.com/',
-    chainId: ChainId.MainnetBeta,
-  },
-  {
-    name: 'mainnet-beta-solana',
-    label: 'mainnet-beta (Solana)',
-    url: 'https://api.mainnet-beta.solana.com',
-    chainId: ChainId.MainnetBeta,
-  },
-  {
-    name: 'mainnet-beta-serum',
-    label: 'mainnet-beta (Serum)',
-    url: 'https://solana-api.projectserum.com/',
     chainId: ChainId.MainnetBeta,
   },
   {
@@ -100,22 +85,15 @@ export function ConnectionProvider({ children }: { children: any }) {
     useLocalStorageState<ENDPOINT_NAME>('network', DEFAULT_ENDPOINT.name);
   const networkParam = searchParams.get('network');
 
-  let maybeEndpoint: Endpoint | undefined = undefined;
-  if (networkParam != null) {
-    maybeEndpoint = ENDPOINTS.find(({ name }) => name === networkParam);
-    if (maybeEndpoint == null && networkParam === LOCALHOST) {
-      // solana-test-validator running on localhost is only used for testing and
-      // not exposed as an endpoint selector option
-      maybeEndpoint = {
-        name: LOCALHOST as ENDPOINT_NAME,
-        label: LOCALHOST,
-        url: 'http://127.0.0.1:8899/',
-        chainId: ChainId.Devnet,
-      };
+  let maybeEndpoint;
+  if (networkParam) {
+    let endpointParam = ENDPOINTS.find(({ name }) => name === networkParam);
+    if (endpointParam) {
+      maybeEndpoint = endpointParam;
     }
   }
 
-  if (networkStorage == null && maybeEndpoint == null) {
+  if (networkStorage && !maybeEndpoint) {
     let endpointStorage = ENDPOINTS.find(({ name }) => name === networkStorage);
     if (endpointStorage) {
       maybeEndpoint = endpointStorage;
@@ -391,6 +369,78 @@ export const sendTransactions = async (
 
   if (sequenceType !== SequenceType.Parallel) {
     await Promise.all(pendingTxns);
+  }
+
+  return signedTxns.length;
+};
+
+export const sendTransactionsWithRecentBlock = async (
+  connection: Connection,
+  wallet: WalletSigner,
+  instructionSet: TransactionInstruction[][],
+  signersSet: Keypair[][],
+  commitment: Commitment = 'singleGossip',
+): Promise<number> => {
+  if (!wallet.publicKey) throw new WalletNotConnectedError();
+
+  const unsignedTxns: Transaction[] = [];
+
+  for (let i = 0; i < instructionSet.length; i++) {
+    const instructions = instructionSet[i];
+    const signers = signersSet[i];
+
+    if (instructions.length === 0) {
+      continue;
+    }
+
+    const block = await connection.getRecentBlockhash(commitment);
+    await sleep(1200);
+
+    const transaction = new Transaction();
+    instructions.forEach(instruction => transaction.add(instruction));
+    transaction.recentBlockhash = block.blockhash;
+    transaction.setSigners(
+      // fee payed by the wallet owner
+      wallet.publicKey,
+      ...signers.map(s => s.publicKey),
+    );
+
+    if (signers.length > 0) {
+      transaction.partialSign(...signers);
+    }
+
+    unsignedTxns.push(transaction);
+  }
+
+  const signedTxns = await wallet.signAllTransactions(unsignedTxns);
+
+  const breakEarlyObject = { breakEarly: false, i: 0 };
+  console.log(
+    'Signed txns length',
+    signedTxns.length,
+    'vs handed in length',
+    instructionSet.length,
+  );
+  for (let i = 0; i < signedTxns.length; i++) {
+    const signedTxnPromise = sendSignedTransaction({
+      connection,
+      signedTransaction: signedTxns[i],
+    });
+
+    signedTxnPromise.catch(() => {
+      breakEarlyObject.breakEarly = true;
+      breakEarlyObject.i = i;
+    });
+
+    try {
+      await signedTxnPromise;
+    } catch (e) {
+      console.log('Caught failure', e);
+      if (breakEarlyObject.breakEarly) {
+        console.log('Died on ', breakEarlyObject.i);
+        return breakEarlyObject.i; // Return the txn we failed on by index
+      }
+    }
   }
 
   return signedTxns.length;